Move configuration into a sub-folder
I think I found a bug with sorbet-coerce that is preventing us from upgrading
sorbet-rails (as newer versions have a dependency on sorbet-coerce). Because
configuration.rb is in the top-level lib/ folder, it is part of the ruby load-path
so require "configuration" will load sorbet-coerce/configuration. This is
breaking for us as we have an engine also called `configuration`, but it now
cannot be loaded without specifying the full path. After monkey-patching bundler
to be more verbose, I see:
```
Loading gem configuration
load paths for configuration/*:
"configuration.rb" => "/Users/kevinmiller/.rvm/gems/ruby-2.6.2/gems/sorbet-coerce-0.2.4/lib"
"configuration/engine.rb" => "/Users/kevinmiller/flexport/engines/configuration/lib"
"configuration/version.rb" => "/Users/kevinmiller/flexport/engines/configuration/lib"
resolved configuration to /Users/kevinmiller/.rvm/gems/ruby-2.6.2/gems/sorbet-coerce-0.2.4/lib/configuration.rb
```
Showing that `require "configuration"` is loading
`/Users/kevinmiller/.rvm/gems/ruby-2.6.2/gems/sorbet-coerce-0.2.4/lib/configuration.rb`

To fix, move configuration to a `public` subfolder. Ideally all of this
gem would be under a subfolder as `require "public/configuration"` is
less clear than `require "sorbet-coerce/public/configuration"`, but that
is a more invasive change (happy to do a followup PR enforcing that if
y'all agree with me)
