+++ lib/public/configuration.rb
